Phase 3a IMPL-3: encoder gain quant uses native (gpQ14, gcMantQ14, gcExp)
Mirror the decoder-side (REF-1 §2 / IMPL-1 / IMPL-2) native g_c
representation on the encoder side so the conjugate-codebook search
and the §A.3.10 commit accumulators no longer collapse the predicted
g_c through a saturated int16 Q12 surrogate.

Changes:

* internal/gain/pow2.go: export Pow2FracQ14 (thin wrapper) so the
  encoder dequant can share the same mantissa table-lookup the decoder
  uses; guarantees encoder ≡ decoder bit-for-bit on (mant, exp).
* internal/gain/seed.go: add SeedDecoder helper to prime a Decoder's
  pastErrors / initialized fields for cross-encoder equivalence tests.
* internal/gainquant/predictor.go:
  - PredictedGcQ12 returns int32 (no int16 saturation) — DIAG-1 §6
    showed g_c0·γ̂_c routinely exceeds the int16 envelope (peaks ≈ 159
    ⇒ Q12 ≈ 651 264); the pre-IMPL-3 saturation biased SearchConjugate.
  - Add predictedLog2GcQ10 (internal log-domain split exposed to the
    dequant path) and DequantGc / Reconstruct helpers that mirror the
    gain.Decoder.Decode (mant Q14, exp int8) pipeline exactly.
* internal/gainquant/conjugate.go: SearchConjugate signature changes
  to gpcPredQ12 int32 and returns γ̂_c Q13 (sum of GBK1[ga][1] +
  GBK2[gb][1]) instead of a saturated ĝc Q12; the cost search uses
  the unsaturated int32 g'c throughout.
* encoder.go fcbStep:
  - Reconstruct the chosen g_c via gainquant.Reconstruct(past, c, ga,
    gb) → (gpQ14, gcMantQ14, gcExp) — same triple a decoder produces.
  - Add mantExpToQ12 to derive a non-saturating int32 Q12 from the
    (mant, exp) pair for the swMemErr / oldExc accumulator multiplies
    (kept as int64 to absorb the exp ≥ 8 codebook envelope).
  - Feed UpdatePastQuaEn the γ̂_c Q13 returned directly by
    SearchConjugate; the pre-IMPL-3 recoverGammaCQ13 inverse-divide
    helper is removed (no consumer).
* internal/gainquant/{predictor,conjugate}_test.go: adapt to the new
  signatures (int32 Q12, γ̂_c return); add gcFromGamma helper for the
  original ĝc-domain assertions.
* internal/gainquant/apply_mantexp_test.go: NEW. Five-case bit-equal
  pin: encoder Reconstruct == decoder gain.Decoder.Decode for matched
  (ga, gb), covering cold-start, voiced-large-gc, low-γ̂_c-floor,
  tiny-single-sample, and zero-energy-guard regimes (REF-1 §2 / IMPL-3
  step C contract).
* internal/gain/legacy_gcq12.go: refresh SCHEDULED FOR REMOVAL header
  to reflect the IMPL-3 landing (no production consumers remain; only
  internal/decoder/*_test.go diagnostic logging probes left for INT-1).
* docs/superpowers/diagnostics/2026-05-04-decoder-amplitude-localization.md:
  append Appendix C with byte-EQ deltas, RoundTrip metrics, and pin
  status post-IMPL-3.

Verification:
* go build ./...                                    : clean
* go vet ./...                                      : clean
* go test ./internal/gainquant/... -count=1         : PASS (incl. new
  TestApply_MantissaExponent — 5 sub-cases bit-equal vs decoder)
* go test ./internal/gain/... ./internal/synth/...  : PASS (zero-alloc
  budget on Decode / BuildExcitation preserved at 0 B/op).
* go test ./... -count=1                            : Pre-existing
  FAIL-DEFERRED set unchanged (TestPhase2fTAME1_ByteEQ,
  TestDiagnostic_SinglePulseChain, 4× PSTdomain
  KnownPSTDomainDifference pins).
* Phase 2c byte-EQ: P1 10.41% / P0 57.22% / P2 11.50% (was 10.79 /
  57.49 / 11.66 — small drift in the expected direction from the now-
  unbiased cost search).
* Phase 2d byte-EQ: S1 5.18% / C1 0.00% / GA1 12.15% / GB1 5.29% /
  S2 4.36% / C2 0.00% / GA2 11.77% / GB2 4.90% (was 5.50 / 0 / 12.15 /
  5.29 / 4.20 / 0 / 11.77 / 4.52).
* Phase 3 RoundTrip SPEECH: B RMS=419 SegSNR=−0.90 dB ; C RMS=5
  SegSNR=0.00 dB (within noise of IMPL-2 baseline).
* Bench (-benchmem): synth.BuildExcitation 230 ns/op 0 B/op ;
  gain.Decode 131 ns/op 0 B/op — zero-alloc preserved.
* I1 (clean-room) risk: none — no external G.729 implementations
  referenced; encoder mirrors decoder math derived in IMPL-1 from
  REF-1 / spec equations.

+++ internal/gainquant/predictor_test.go
@@ -33,9 +33,9 @@ func TestPredictedGcQ12_ColdStartFourPulses(t *testing.T) {
 
 	got := PredictedGcQ12(&past, &c)
 
-	const wantApprox int16 = 22878 // Q12 (5.585 × 2^12)
+	const wantApprox int32 = 22878 // Q12 (5.585 × 2^12)
 	const tol = 230                // ≈ 1%
-	diff := int32(got) - int32(wantApprox)
+	diff := got - wantApprox
 	if diff < -tol || diff > tol {
 		t.Fatalf("PredictedGcQ12 cold-start 4 pulses = %d (Q12), want %d ±%d",
 			got, wantApprox, tol)
